remove(ng): move custom marker to javascript folder
