+++ app/javascript/javascript/customMarker.js
